renaming netstack/type.h to event/buffer.h
